Unsynced merge week (#8)
* crud ops on file handling completed -- to fix file size not showing up

* updated delete route to cascade delete files and comments

* build comment form, handle client storage, && build view client comments

* update view single track page && build comments section on page dynamically

* fix file size not being added to database, modified pages to show file size and total storage data | cleaned console logs

* restructure project for easier access to components

* maintenance, fix reset state/routing issue, add loading state to prevent component render errors.

* handle profile page to allow user crud => implement toast to login and logout user

* fix csrf issues

* finalize crud functions between frontend and backend

* add client view and update project share link

* handled resetting form fields for dialog boxes => ensured no memory leaks while checking for state change

* implemented toast for Edit project form and File uploading

* added download link to get file(s) -- will update to download all files at once

* add toast to upload and delete file

* added payment processing popup

* Fix/reinit repo (#7)

* convert repo to monorepo

* fix(csrf): fix invalid csrf token issue => now storing sessions in mongodb

+++ frontend/src/pages/user/projects/ViewProjectsPage.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { Eye, FolderKanban, Inbox, Trash2 } from "lucide-react";
+import { Eye, FolderKanban, Trash2 } from "lucide-react";
 import {
   Table,
   TableBody,
@@ -23,34 +23,62 @@ import {
   selectAllProjects
 } from "@/features/projects/projectsSlice";
 import { formatDistanceToNow, parseISO } from "date-fns";
-import { Link, useNavigate } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import NewProjectFormPopup from "@/components/popups/NewProjectForm";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger
+} from "@/components/ui/alert-dialog";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
 
 export default function ViewProjectsPage() {
   const dispatch = useDispatch();
   const navigate = useNavigate();
   const projects = useSelector(selectAllProjects);
-  const formattedProjects = projects.map((project) => {
-    const date = parseISO(project.createdAt);
-    const timePeriod = formatDistanceToNow(date);
-    const timeAgo = `${timePeriod} ago`;
+  const formattedProjects = projects
+    .slice()
+    .sort((a, b) => b.createdAt.localeCompare(a.createdAt))
+    .map((project) => {
+      const date = parseISO(project.createdAt);
+      const timePeriod = formatDistanceToNow(date);
+      const timeAgo = `${timePeriod} ago`;
 
-    return {
-      id: project._id,
-      title: project.title,
-      createdAt: timeAgo,
-      status: project.status
-    };
-  });
+      return {
+        id: project._id,
+        title: project.title,
+        createdAt: timeAgo,
+        status: project.status
+      };
+    });
 
-  const [selected, setSelected] = useState([]);
+  const [deleteInput, setDeleteInput] = useState("");
+  const [deleteDialogOpen, setDeleteDialogOpen] = useState(false);
 
   useEffect(() => {
     dispatch(getProjects());
   }, [dispatch]);
+
+  useEffect(() => {
+    if (!deleteDialogOpen) {
+      setDeleteInput("");
+    }
+  }, [deleteDialogOpen]);
+
+  const handleDeleteInput = (e) => {
+    setDeleteInput(e.target.value);
+  };
   const handleDeleteProject = async (projectId) => {
     await dispatch(deleteProject(projectId));
   };
+
   return (
     <section className="m-8">
       <div className="flex items-center justify-between mb-4">
@@ -103,12 +131,49 @@ export default function ViewProjectsPage() {
                   <TooltipProvider>
                     <Tooltip>
                       <TooltipTrigger asChild>
-                        <Button
-                          variant="destructive"
-                          onClick={() => handleDeleteProject(project.id)}
+                        <AlertDialog
+                          open={deleteDialogOpen}
+                          onOpenChange={setDeleteDialogOpen}
                         >
-                          <Trash2 />
-                        </Button>
+                          <AlertDialogTrigger asChild>
+                            <Button variant="destructive">
+                              <Trash2 />
+                            </Button>
+                          </AlertDialogTrigger>
+                          <AlertDialogContent>
+                            <AlertDialogHeader>
+                              <AlertDialogTitle>
+                                Confirm Delete Project?
+                              </AlertDialogTitle>
+                              <AlertDialogDescription>
+                                This action cannot be undone. This will
+                                permanently delete your account and remove your
+                                data from our servers.
+                              </AlertDialogDescription>
+                            </AlertDialogHeader>
+                            <div>
+                              <Label>
+                                Type <em>'{project.title}'</em> to confirm
+                              </Label>
+                              <Input
+                                type="text"
+                                value={deleteInput}
+                                onChange={handleDeleteInput}
+                              />
+                            </div>
+                            <AlertDialogFooter>
+                              <AlertDialogCancel>
+                                Take me back
+                              </AlertDialogCancel>
+                              <AlertDialogAction
+                                onClick={() => handleDeleteProject(project.id)}
+                                disabled={deleteInput !== project.title}
+                              >
+                                Yes, delete this project
+                              </AlertDialogAction>
+                            </AlertDialogFooter>
+                          </AlertDialogContent>
+                        </AlertDialog>
                       </TooltipTrigger>
                       <TooltipContent>
                         <p>Delete project</p>
